baselines/mainmethod_gemini/requirements.txt
google-generativeai
PyYAML
tenacity
tqdm
scikit-learn
